feat(View): Improve bottom view

+++ app/src/main/java/com/bei/andrii/forexcalculator/activity/MainActivity.java
@@ -4,7 +4,7 @@ public interface MainActivity {
 
     void showResultForStep(int mPPrice, int mPStop, int mPTake);
 
-    void showResultAmount(float mAmountEnter, double mExpectedProfit, double mExpectedLesion, float mExpectedProfitPercent, float mExpectedLesionPercent);
+    void showResultAmount(float mAmountEnter, float mExpectedProfit, float mExpectedLesion, float mExpectedProfitPercent, float mExpectedLesionPercent);
 
     void showResultComment(int mComment);
